% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hannibalRutils.R
\name{convert.cprTodate}
\alias{convert.cprTodate}
\title{Convert Cpr to date}
\usage{
convert.cprTodate(x)
}
\arguments{
\item{x}{A character string containing a cpr number}
}
\value{
Returns the birthday of a person given a cpr number
}
\description{
Convert Cpr to date
}
\examples{

cpr <- 0203592008

convert.cprTodate(cpr)

}
