% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hannibalRutils.R
\name{convertToCPR}
\alias{convertToCPR}
\title{converts numeric CPR numbers to characters and adds a leading 0 to danish CPR number of only 7 digits.}
\usage{
convertToCPR(vector)
}
\arguments{
\item{vector}{a vector containing numeric cpr numbers}
}
\value{
A vector containing danish cpr number in character form adding a possible leading 0.
}
\description{
converts numeric CPR numbers to characters and adds a leading 0 to danish CPR number of only 7 digits.
}
\examples{

vec <- c(9076342, 3102314, 24032789)

convertToCPR(vec)

}
