% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hannibalRutils.R
\name{dropcols}
\alias{dropcols}
\title{Removes columns from data frames}
\usage{
dropcols(data, colnames)
}
\arguments{
\item{data}{A data frame containing the column to remove}

\item{colnames}{The name of the columnn that is to be removed}
}
\value{
Nothing, however runs code to remove volumns from a data frame
}
\description{
Removes columns from data frames
}
\examples{

col1 <- c("Area1", "Area2", "Area3")

col2 <- c(10, 20, 30)

df <- data.frame(Area = col1, Size = col2)


dropcols(df, "Area")




}
