% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hannibalRutils.R
\name{extractVectorCode}
\alias{extractVectorCode}
\title{Extract code to hard code creation of a given vector}
\usage{
extractVectorCode(vector)
}
\arguments{
\item{vector}{A vector for which is wanted to be hard coded.}
}
\value{
Returns a text that represent a code snippet to hard code creation of the given vecor
}
\description{
Extract code to hard code creation of a given vector
}
\examples{

vec <- c("Word1 word2", "Word3 word4")

extractVectorCode(vec)

}
