% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hannibalRutils.R
\name{factorize}
\alias{factorize}
\title{Factorize a column in a dataframe}
\usage{
factorize(x)
}
\arguments{
\item{x}{A vector to convert to vector, normally a column in a dataframe}
}
\value{
Returns nothing bu prints the code used for factorizing the specific column
}
\description{
Factorize a column in a dataframe
}
\examples{

col1 <- c("Area1", "Area2", "Area3")

col2 <- c(10, 20, 30)

df <- data.frame(Area = col1, Size = col2)


factorize(df$Area)


}
