% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hannibalRutils.R
\name{SectionComment}
\alias{SectionComment}
\title{Create a section comment for easy division and overview of scripts}
\usage{
SectionComment(text = "Example text", width = 80)
}
\arguments{
\item{text}{The text to be displayed, must be at least two characters shorter than width}

\item{width}{How wide should the comment be}
}
\value{
A text containing the comment
}
\description{
Create a section comment for easy division and overview of scripts
}
\examples{

SectionComment("Section 1")


}
