% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hannibalRutils.R
\name{SpaceReplace}
\alias{SpaceReplace}
\title{Replace all white spaces in a character vector with underscores}
\usage{
SpaceReplace(vector)
}
\arguments{
\item{vector}{A character vector containing text where spaces needs to be replaced}
}
\value{
Returns the provided vector with the white spaces replaced with underscores
}
\description{
Replace all white spaces in a character vector with underscores
}
\examples{
vec <- c("Word1 word2", "Word3 word4")

SpaceReplace(vec)


}
